+++ projects/plugins/backup/src/js/components/Admin.js
@@ -65,12 +65,10 @@ const Admin = () => {
 								'Get peace of mind knowing your work will be saved, add backups today.',
 								'jetpack-backup'
 							) }
-							<br />
-							{ __( 'Choose from real time or daily backups.', 'jetpack-backup' ) }
 						</p>
 						<a
 							class="button"
-							href={ getRedirectUrl( 'backup-plugin-upgrade-daily', { site: domain } ) }
+							href={ getRedirectUrl( 'backup-plugin-upgrade-10gb', { site: domain } ) }
 							target="_blank"
 							rel="noreferrer"
 						>