+++ projects/plugins/jetpack/_inc/client/main.jsx
@@ -6,6 +6,7 @@ import { connect } from 'react-redux';
 import { withRouter, Prompt } from 'react-router-dom';
 import { __, sprintf } from '@wordpress/i18n';
 import { getRedirectUrl } from '@automattic/jetpack-components';
+import { PartnerCouponRedeem } from '@automattic/jetpack-partner-coupon';
 import { ConnectScreen, CONNECTION_STORE_ID } from '@automattic/jetpack-connection';
 import { Dashicon } from '@wordpress/components';
 import { withDispatch } from '@wordpress/data';
@@ -18,6 +19,11 @@ import Navigation from 'components/navigation';
 import NavigationSettings from 'components/navigation-settings';
 import SearchableSettings from 'settings/index.jsx';
 import {
+	updateLicensingActivationNoticeDismiss as updateLicensingActivationNoticeDismissAction,
+	updateUserLicensesCounts as updateUserLicensesCountsAction,
+} from 'state/licensing';
+import { fetchModules as fetchModulesAction } from 'state/modules';
+import {
 	getSiteConnectionStatus,
 	isCurrentUserLinked,
 	isSiteConnected,
@@ -29,6 +35,8 @@ import {
 	getConnectingUserFeatureLabel,
 	getConnectionStatus,
 	hasConnectedOwner,
+	getHasSeenWCConnectionModal,
+	setHasSeenWCConnectionModal,
 } from 'state/connection';
 import {
 	setInitialState,
@@ -43,9 +51,19 @@ import {
 	getTracksUserData,
 	showRecommendations,
 	getPluginBaseUrl,
+	getPartnerCoupon,
 	isWoASite,
+	isWooCommerceActive,
 } from 'state/initial-state';
-import { areThereUnsavedSettings, clearUnsavedSettingsFlag } from 'state/settings';
+import {
+	fetchSiteData as fetchSiteDataAction,
+	fetchSitePurchases as fetchSitePurchasesAction,
+} from 'state/site';
+import {
+	areThereUnsavedSettings,
+	clearUnsavedSettingsFlag,
+	fetchSettings as fetchSettingsAction,
+} from 'state/settings';
 import { getSearchTerm } from 'state/search';
 import { Recommendations } from 'recommendations';
 import ProductDescriptions from 'product-descriptions';
@@ -65,6 +83,9 @@ import QueryRewindStatus from 'components/data/query-rewind-status';
 import { getRewindStatus } from 'state/rewind';
 import ReconnectModal from 'components/reconnect-modal';
 import { createInterpolateElement } from '@wordpress/element';
+import { imagePath } from 'constants/urls';
+import { ActivationScreen } from '@automattic/jetpack-licensing';
+import ContextualizedConnection from 'components/contextualized-connection';
 
 const recommendationsRoutes = [
 	'/recommendations',
@@ -95,6 +116,7 @@ class Main extends React.Component {
 	constructor( props ) {
 		super( props );
 		this.closeReconnectModal = this.closeReconnectModal.bind( this );
+		this.onLicenseActivationSuccess = this.onLicenseActivationSuccess.bind( this );
 	}
 
 	UNSAFE_componentWillMount() {
@@ -102,8 +124,6 @@ class Main extends React.Component {
 		restApi.setApiRoot( this.props.apiRoot );
 		restApi.setApiNonce( this.props.apiNonce );
 		this.initializeAnalytics();
-
-		this.props.setConnectionStatus( this.props.connectionStatus );
 
 		// Handles refresh, closing and navigating away from Jetpack's Admin Page
 		// beforeunload can not handle confirm calls in most of the browsers, so just clean up the flag.
@@ -124,6 +144,22 @@ class Main extends React.Component {
 		const fullScreenContainer = jQuery( '.jp-connect-full__container' );
 		if ( connectReactContainer && fullScreenContainer.length > 0 ) {
 			fullScreenContainer.prependTo( connectReactContainer );
+		}
+
+		/* Redirects the user to the Woo Connection/Welcome Screen if:
+		 * 1. They have Woo installed and active AND
+		 * 2. they have never seen this screen before AND
+		 * 3. they have the right permissions.
+		 */
+		if (
+			this.props.isWooCommerceActive &&
+			! this.props.hasSeenWCConnectionModal &&
+			this.props.userCanManageModules
+		) {
+			this.props.history.replace( {
+				pathname: '/woo-setup',
+				state: { previousPath: this.props.location.pathname },
+			} );
 		}
 	}
 
@@ -198,6 +234,83 @@ class Main extends React.Component {
 	}
 
 	renderMainContent = route => {
+		if ( this.shouldShowWooConnectionScreen() ) {
+			const previousPath = this.props.location.state?.previousPath;
+			const redirectTo =
+				previousPath && previousPath !== '/woo-setup' ? `#${ previousPath }` : '#/dashboard';
+
+			return (
+				<ContextualizedConnection
+					apiNonce={ this.props.apiNonce }
+					registrationNonce={ this.props.registrationNonce }
+					apiRoot={ this.props.apiRoot }
+					title={ __(
+						'Welcome to Jetpack! Security, Growth, & Performance tools for WordPress businesses',
+						'jetpack'
+					) }
+					logo={
+						<img
+							src={ imagePath + '/jetpack-woocommerce-logo.svg' }
+							alt={ __( 'Jetpack and WooCommerce', 'jetpack' ) }
+						/>
+					}
+					buttonLabel={ __( 'Set up Jetpack', 'jetpack' ) }
+					redirectUri="admin.php?page=jetpack"
+					redirectTo={ redirectTo }
+					from={ this.props.location.pathname }
+					isSiteConnected={ this.props.isSiteConnected }
+					setHasSeenWCConnectionModal={ this.props.setHasSeenWCConnectionModal }
+				>
+					<p>
+						{ __(
+							'Jetpack is the perfect companion plugin for WooCommerce - made by WordPress experts to make your store faster, safer, and to help grow your business.',
+							'jetpack'
+						) }
+					</p>
+				</ContextualizedConnection>
+			);
+		}
+
+		/*
+		 * Show "Partner Coupon Redeem" screen instead of regular main content/pre-connection.
+		 */
+		if ( this.props.partnerCoupon ) {
+			const forceShow = new URLSearchParams( window.location.search ).get( 'showCouponRedemption' );
+
+			/*
+			 * There are two conditions (groups of conditions, really) where we would want to
+			 * show the partner coupon redeem screen:
+			 *
+			 * 1. The site is not yet connected to WPCOM, but has the jetpack_partner_coupon
+			 * option set in the database (this.props.partnerCoupon in redux). This is likely a
+			 * partner-user who has just arrived here from a CTA within a partner's dashboard
+			 * or other ecosystem.
+			 *
+			 * 2. The site is already connected to WPCOM, but the jetpack_partner_coupon option
+			 * is still set in the database. This means the user connected their site, but never
+			 * redeemed the coupon. If this is the case, we don't want to override the dashboard
+			 * or at a glance pages with the redemption screen. Instead, we'll catch a URL
+			 * parameter that JITMs will set (showCouponRedemption=true), and show the screen only
+			 * when the user came from a a JITM.
+			 */
+			if ( ! this.props.isSiteConnected || forceShow ) {
+				return (
+					<PartnerCouponRedeem
+						apiNonce={ this.props.apiNonce }
+						registrationNonce={ this.props.registrationNonce }
+						apiRoot={ this.props.apiRoot }
+						images={ [ '/images/connect-right-partner-backup.png' ] }
+						assetBaseUrl={ this.props.pluginBaseUrl }
+						connectionStatus={ this.props.connectionStatus }
+						partnerCoupon={ this.props.partnerCoupon }
+						siteRawUrl={ this.props.siteRawUrl }
+						tracksUserData={ !! this.props.tracksUserData }
+						analytics={ analytics }
+					/>
+				);
+			}
+		}
+
 		if (
 			this.isUserConnectScreen() &&
 			( this.props.userCanManageModules || this.props.hasConnectedOwner )
@@ -309,6 +422,7 @@ class Main extends React.Component {
 			case '/reconnect':
 			case '/disconnect':
 			case '/connect-user':
+			case '/woo-setup':
 			case '/setup':
 				pageComponent = (
 					<AtAGlance
@@ -349,6 +463,18 @@ class Main extends React.Component {
 						searchTerm={ this.props.searchTerm }
 						rewindStatus={ this.props.rewindStatus }
 						userCanManageModules={ this.props.userCanManageModules }
+					/>
+				);
+				break;
+			case '/license/activation':
+				navComponent = null;
+				pageComponent = (
+					<ActivationScreen
+						assetBaseUrl={ this.props.pluginBaseUrl }
+						lockImage="/images/jetpack-license-activation-with-lock.png"
+						siteRawUrl={ this.props.siteRawUrl }
+						successImage="/images/jetpack-license-activation-with-success.png"
+						onActivationSuccess={ this.onLicenseActivationSuccess }
 					/>
 				);
 				break;
@@ -406,12 +532,20 @@ class Main extends React.Component {
 
 	shouldShowAppsCard() {
 		// Only show on the dashboard
-		return this.props.isSiteConnected && dashboardRoutes.includes( this.props.location.pathname );
+		return (
+			this.props.isSiteConnected &&
+			! this.shouldShowWooConnectionScreen() &&
+			dashboardRoutes.includes( this.props.location.pathname )
+		);
 	}
 
 	shouldShowSupportCard() {
 		// Only show on the dashboard
-		return this.props.isSiteConnected && dashboardRoutes.includes( this.props.location.pathname );
+		return (
+			this.props.isSiteConnected &&
+			! this.shouldShowWooConnectionScreen() &&
+			dashboardRoutes.includes( this.props.location.pathname )
+		);
 	}
 
 	shouldShowRewindStatus() {
@@ -467,6 +601,15 @@ class Main extends React.Component {
 	}
 
 	/**
+	 * Checks whether we should show the Woo Connection screen page.
+	 *
+	 * @returns {boolean} Whether we should show the Woo connection screen page.
+	 */
+	shouldShowWooConnectionScreen() {
+		return '/woo-setup' === this.props.location.pathname;
+	}
+
+	/**
 	 * Check if the user connection has been triggered.
 	 *
 	 * @returns {boolean} Whether the user connection has been triggered.
@@ -484,12 +627,40 @@ class Main extends React.Component {
 	}
 
 	/**
+	 * Checks if this is a licensing screen page.
+	 *
+	 * @returns {boolean} Whether this is a licensing screen page.
+	 */
+	isLicensingScreen() {
+		return this.props.location.pathname.startsWith( '/license' );
+	}
+
+	/**
 	 * Check if the connection flow should get triggered automatically.
 	 *
 	 * @returns {boolean} Whether to trigger the connection flow automatically.
 	 */
 	shouldAutoTriggerConnection() {
 		return this.props.location.pathname.startsWith( '/setup' );
+	}
+
+	/**
+	 * Fires after a user(not partner) product license key has been sucessfully activated.
+	 */
+	onLicenseActivationSuccess() {
+		// First update state.jetpack.licensing.userCounts before dismissing the license activation notice.
+		this.props.updateUserLicensesCounts().then( () => {
+			// Manually dismiss the userLicenseActivationNotice.
+			this.props.updateLicensingActivationNoticeDismiss();
+		} );
+		// Update site plan.
+		this.props.fetchSiteData();
+		// Update site products.
+		this.props.fetchSitePurchases();
+		// Update site modules (search, wordads, google-analytics, etc.)
+		this.props.fetchModules();
+		// Update site settings (i.e. search, instant search, etc.)
+		this.props.fetchSettings();
 	}
 
 	render() {
@@ -501,6 +672,10 @@ class Main extends React.Component {
 
 		if ( this.isUserConnectScreen() ) {
 			jpClasses.push( 'jp-user-connect-screen' );
+		}
+
+		if ( this.isLicensingScreen() ) {
+			jpClasses.push( 'jp-licensing-screen' );
 		}
 
 		return (
@@ -556,6 +731,9 @@ export default connect(
 			connectUrl: getConnectUrl( state ),
 			connectingUserFeatureLabel: getConnectingUserFeatureLabel( state ),
 			isWoaSite: isWoASite( state ),
+			isWooCommerceActive: isWooCommerceActive( state ),
+			hasSeenWCConnectionModal: getHasSeenWCConnectionModal( state ),
+			partnerCoupon: getPartnerCoupon( state ),
 		};
 	},
 	dispatch => ( {
@@ -568,17 +746,36 @@ export default connect(
 		reconnectSite: () => {
 			return dispatch( reconnectSite() );
 		},
+		setHasSeenWCConnectionModal: () => {
+			return dispatch( setHasSeenWCConnectionModal() );
+		},
 		resetConnectUser: () => {
 			return dispatch( resetConnectUser() );
+		},
+		updateLicensingActivationNoticeDismiss: () => {
+			return dispatch( updateLicensingActivationNoticeDismissAction() );
+		},
+		updateUserLicensesCounts: () => {
+			return dispatch( updateUserLicensesCountsAction() );
+		},
+		fetchSiteData: () => {
+			return dispatch( fetchSiteDataAction() );
+		},
+		fetchSitePurchases: () => {
+			return dispatch( fetchSitePurchasesAction() );
+		},
+		fetchModules: () => {
+			return dispatch( fetchModulesAction() );
+		},
+		fetchSettings: () => {
+			return dispatch( fetchSettingsAction() );
 		},
 	} )
 )(
 	withDispatch( dispatch => {
 		return {
 			setConnectionStatus: connectionStatus => {
-				dispatch( CONNECTION_STORE_ID ).startResolution( 'getConnectionStatus', [] );
 				dispatch( CONNECTION_STORE_ID ).setConnectionStatus( connectionStatus );
-				dispatch( CONNECTION_STORE_ID ).finishResolution( 'getConnectionStatus', [] );
 			},
 		};
 	} )( withRouter( Main ) )