+++ projects/packages/options/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.0] - 2021-11-30
+### Added
+- Added has_seen_wc_connection_modal option to the list of available options
+- Add new Jetpack plugin partner coupon options
+
+## [1.13.5] - 2021-11-16
+### Added
+- Added Jetpack Option 'licensing_activation_notice_dismiss'.
+
 ## [1.13.4] - 2021-10-13
 ### Changed
 - Updated package dependencies.
@@ -128,6 +137,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[1.14.0]: https://github.com/Automattic/jetpack-options/compare/v1.13.5...v1.14.0
+[1.13.5]: https://github.com/Automattic/jetpack-options/compare/v1.13.4...v1.13.5
 [1.13.4]: https://github.com/Automattic/jetpack-options/compare/v1.13.3...v1.13.4
 [1.13.3]: https://github.com/Automattic/jetpack-options/compare/v1.13.2...v1.13.3
 [1.13.2]: https://github.com/Automattic/jetpack-options/compare/v1.13.1...v1.13.2